I've changed the code to use a "level-pool" volume reservoir boundary condition instead of the previous method that relied on the tunnel volume as computed from the HPGs.
I've run a comparison between the "level-pool" reservoir boundary condition and the previous method (tunnel volume).  The maximum relative percent error is 286%.  It is important to keep in mind that high % error values occur only for a short time at the beginning of a storm event when the true volume has not exceeded the tunnel ponded storage volume.  Based on the data, I believe that the "level-pool" volume method is sufficiently accurate for our purposes.

You can find the data here:

   \\cee-hsl-tarp-fs\tarp\Software\UIUC\Hydraulic\ICAP\reservoir-bc